Datasets/evalplus/Java/src/test/java/com/example/HumanEval_129Test.java
package com.example;
import org.junit.Test;
import static org.junit.Assert.*;
import java.util.Arrays;
import java.util.List;

public class HumanEval_129Test {

    @Test
    public void test_0() {
        HumanEval_129 tester = new HumanEval_129();
        int[][] grid = {{1, 2, 3}, {4, 5, 6}, {7, 8, 9}};
        int k = 3;
        List<Integer> result = Arrays.asList(1, 2, 1);
        assertEquals(result, tester.minPath(grid, k));
    }

    @Test
    public void test_1() {
        HumanEval_129 tester = new HumanEval_129();
        int[][] grid = {{5, 9, 3}, {4, 1, 6}, {7, 8, 2}};
        int k = 1;
        List<Integer> result = Arrays.asList(1);
        assertEquals(result, tester.minPath(grid, k));
    }

    @Test
    public void test_2() {
        HumanEval_129 tester = new HumanEval_129();
        int[][] grid = {{1, 2, 3, 4}, {5, 6, 7, 8}, {9, 10, 11, 12}, {13, 14, 15, 16}};
        int k = 4;
        List<Integer> result = Arrays.asList(1, 2, 1, 2);
        assertEquals(result, tester.minPath(grid, k));
    }

    @Test
    public void test_3() {
        HumanEval_129 tester = new HumanEval_129();
        int[][] grid = {{6, 4, 13, 10}, {5, 7, 12, 1}, {3, 16, 11, 15}, {8, 14, 9, 2}};
        int k = 7;
        List<Integer> result = Arrays.asList(1, 10, 1, 10, 1, 10, 1);
        assertEquals(result, tester.minPath(grid, k));
    }

    // Continue with the rest of the tests in the same manner...

}
// End of Test
